Standard icon path for linux
I think it would be better, the main icon is placed in the hicolor icon which is the standard icon for linux.

`icons/hicolor/48x48/apps`

But it would be even better if a scalable icon was provided, then placed in `icons/hicolor/scalable/apps`. Unfortunately, I didn't find the svg icon.